requirements.txt
-e .
setuptools
django
pytest-django
django-redis-cache
hiredis
pytest-xdist
tox
wheel
twine
flake8

